Added separate utility files for print_array and check_if_equal. Commented testbench.' Added block comments to header files
Further comments:

Add author, email, references, etc. to testbench and benchmark files
Where is the input data for the filter?  Or is there input data?
Please comment the elliptic filter code
