+++ elliptic_tb/elliptic_tb.c
@@ -1,27 +1,44 @@
+/**
+    Elliptic filter testbench.
 
-#include "file_selector.h"
+    Reference: (https://en.wikipedia.org/wiki/Elliptic_filter)
 
-#if ELLIPTIC_TB
+    @file elliptic_tb.c
+    @author
+    @email
+    @date
+*/
 
 #include "elliptic.h"
-#include <stdio.h>
+#include "util.h"
 
+/************************************************
+ *  constants
+ ***********************************************/
 
 #define PADDING 5
 #define DATA_LENGTH 256
 
 
+/************************************************
+ *  main
+ ***********************************************/
+
 int main(void)
 {
+    // filter coefficients
     float a[] = { 1, 1.6141607939064317, -0.9031492413866293, 1, 1.6789212089700198, -0.8298796467544193, 1, 1.7402021652716963, -0.7739286551279295, 1, 1.5944666529140512, -0.9692173129574591 };
     float b[] = { 1, -0.9852674718763040, 1, 1, -0.3678140918675017, 1, 1, 1.3541711838514376, 1, 1, -1.1713815107285585, 1 };
-    float y[DATA_LENGTH];
+    // filter gains
     float gain_factor = 0.085667864748512997929985601786029292271 * 0.071984525012505143348739977682271273807 * 0.06414311970061792322894689277745783329 * 0.060572179183183680328816933524649357423;
-    
-    float* a_ptr = &a;
-    filter_response(&y, a_ptr, &b, gain_factor, PADDING, DATA_LENGTH);
+    // filter response
+    float res[DATA_LENGTH];
 
-    float check_data[] = { 0.000024, 0.000131, 0.000449, 0.001190, 0.002645, 0.005129, 0.008886, 0.013952, 0.020026, 0.026398, 0.031986, 0.035517, 0.035822, 0.032192, 0.024681, 0.014245,
+    // compute filter response
+    elliptic(&res[0], &a[0], &b[0], gain_factor, PADDING, DATA_LENGTH);
+
+    // expected (true) result
+    float true_res[] = { 0.000024, 0.000131, 0.000449, 0.001190, 0.002645, 0.005129, 0.008886, 0.013952, 0.020026, 0.026398, 0.031986, 0.035517, 0.035822, 0.032192, 0.024681, 0.014245,
                             0.002635, -0.007959, -0.015441, -0.018380, -0.016428, -0.010463, -0.002369, 0.005483, 0.010903, 0.012515, 0.010130, 0.004754, -0.001765, -0.007350, -0.010356,
                             -0.010068, -0.006876, -0.002082, 0.002584, 0.005568, 0.006004, 0.003966, 0.000388, -0.003310, -0.005744, -0.006053, -0.004177, -0.000840, 0.002742, 0.005309,
                             0.006008, 0.004672, 0.001851, -0.001417, -0.003993, -0.005038, -0.004294, -0.002140, 0.000582, 0.002886, 0.004007, 0.003649, 0.002062, -0.000084, -0.001975,
@@ -39,9 +56,10 @@ int main(void)
                             0.000216, 0.000183, 0.000082, -0.000046, -0.000153, -0.000200, -0.000170, -0.000077, 0.000041, 0.000141, 0.000185, 0.000158, 0.000073, -0.000037, -0.000129,
                             -0.000171, -0.000146, -0.000068, 0.000033, 0.000119, 0.000158, 0.000136, 0.000064, -0.000030, -0.000109, -0.000146, -0.000126, -0.000060, 0.000027, 0.000101 };
 
+    // test for pointwise equality with some error tolerance
+    //      abs(res[i] - true_res[i]) < epsilon
     float epsilon = 0.000002;
-
-    int pass = check_if_equal(y, check_data, DATA_LENGTH, epsilon);
+    int pass = check_if_equal(res, true_res, DATA_LENGTH, epsilon);
 
     if (pass) {
         printf("Test Passed!");
@@ -52,5 +70,3 @@ int main(void)
 
     return 0;
 }
-
-#endif //ELLIPTIC_TB
